+++ Runtime/Constant/FalseValidator.cs
@@ -1,6 +1,6 @@
 ﻿namespace dGameBoy101b.Validators
 {
-	public sealed class FalseValidator : ValidatorBase
+	public sealed class FalseValidator : Validator
 	{
 		public override bool CheckValidity()
 		{