+++ Runtime/Validator.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 namespace dGameBoy101b.Validators
 {
-	public abstract class ValidatorBase : MonoBehaviour
+	public abstract class Validator : MonoBehaviour
 	{
 		public abstract bool CheckValidity();
 	}